requirements.txt
fastapi[standard]
gspread 
google-auth